backend/requirements.txt
# FastAPI and dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Modbus TCP communication library
pymodbus>=3.0.0

# Redis for data storage
redis>=5.0.0

# Configuration management
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# Async support
asyncio